Refactor/remove unused Rake tasks files
