--- rake-tasks/browsers.rb
@@ -1,1 +1,0 @@
-require 'rake-tasks/checks'
