--- rake-tasks/c.rb
@@ -1,6 +1,0 @@
-# C related tasks
-
-require 'rake-tasks/files'
-
-# Location of all new methods
-require 'rake-tasks/selenium_rake/c_tasks'
