--- rake-tasks/checks.rb
@@ -1,2 +1,0 @@
-require 'rake-tasks/selenium_rake/checks'
-# All moved to selenium_rake/checks
