--- rake-tasks/crazy_fun.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-require 'rake-tasks/crazy_fun/main'
-require 'rake-tasks/selenium_rake/detonating_handler'
-require 'rake-tasks/selenium_rake/crazy_fun'
