--- rake-tasks/ie_code_generator.rb
@@ -1,4 +1,0 @@
-require 'rake-tasks/selenium_rake/ie_code_generator'
-require 'rake-tasks/selenium_rake/java_formatter'
-require 'rake-tasks/selenium_rake/cpp_formatter'
-require 'rake-tasks/selenium_rake/type_definitions_generator'
