+++ rake_tasks/crazy_fun/mappings/export.rb
@@ -1,4 +1,4 @@
-require 'rake-tasks/crazy_fun/mappings/common'
+require 'rake_tasks/crazy_fun/mappings/common'
 
 class ExportMappings
   def add_all(fun)
