+++ rake_tasks/crazy_fun/mappings/ruby.rb
