+++ rake_tasks/selenium_rake/base_generator.rb
