+++ rake_tasks/selenium_rake/cpp_formatter.rb
