+++ rake_tasks/selenium_rake/crazy_fun.rb
