+++ rake_tasks/selenium_rake/detonating_handler.rb
