+++ rake_tasks/selenium_rake/java_formatter.rb
